Convert DommiksCraftingStore plugin from kts to kt
